Revert "Use file_writer_config instead of config to get text_dir for Text() (#220)"
This reverts commit 7f2876991d64c436ae9bd4c72f3929d169c4501a.
