+++ manim/utils/file_ops.py
@@ -1,6 +1,4 @@
 import os
-import subprocess as sp
-import platform
 import numpy as np
 import time
 
@@ -39,18 +37,3 @@ def modify_atime(file_path):
         The path of the file. 
     """
     os.utime(file_path, times=(time.time(),os.path.getmtime(file_path)))
-
-def open_file(file_path):
-    current_os = platform.system()
-    if current_os == "Windows":
-        os.startfile(file_path)
-    else:
-        commands = []
-        if current_os == "Linux":
-            commands.append("xdg-open")
-        elif current_os.startswith("CYGWIN"):
-            commands.append("cygstart")
-        else:  # Assume macOS
-            commands.append("open")
-        commands.append(file_path)
-        sp.Popen(commands)
