+++ setup.py
@@ -11,6 +11,7 @@ setup(
         "console_scripts": [
             "manim=manim.__main__:main",
             "manimcm=manim.__main__:main",
+            "manim-cfg=manim.utils.cfgwriter:main",
         ]
     },
     install_requires=[
