+++ tests/test_cli/manim.cfg
@@ -1,7 +1,6 @@
 [CLI]
 movie_file_extension = .mp4
 write_to_movie = True
-sound =  True
 # write_all = False
 save_last_frame = True
 # save_pngs = False
