--- tests/test_cli/write_cfg_sbcmd_input.txt
@@ -1,26 +1,0 @@
-
-
-
-
-
-
-
-
-False
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
